+++ packages/camera/camera_avfoundation/example/integration_test/camera_test.dart
@@ -215,9 +215,13 @@ void main() {
     await controller.prepareForVideoRecording();
 
     await controller.startVideoRecording();
+<<<<<<< HEAD
     sleep(const Duration(milliseconds: 500));
     await controller.setDescription(cameras[1]);
     sleep(const Duration(milliseconds: 500));
+=======
+    await controller.setDescription(cameras[1]);
+>>>>>>> main
 
     expect(controller.description, cameras[1]);
   });
@@ -236,9 +240,13 @@ void main() {
     );
 
     await controller.initialize();
+<<<<<<< HEAD
     sleep(const Duration(milliseconds: 500));
     await controller.setDescription(cameras[1]);
     sleep(const Duration(milliseconds: 500));
+=======
+    await controller.setDescription(cameras[1]);
+>>>>>>> main
 
     expect(controller.description, cameras[1]);
   });