+++ packages/google_sign_in/google_sign_in/example/lib/main.dart
@@ -30,7 +30,7 @@ void main() {
 }
 
 class SignInDemo extends StatefulWidget {
-  const SignInDemo({Key? key}) : super(key: key);
+  const SignInDemo({super.key});
 
   @override
   State createState() => SignInDemoState();