+++ packages/image_picker/image_picker_android/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerDelegate.java
@@ -15,13 +15,17 @@ import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.MediaStore;
+import androidx.activity.result.PickVisualMediaRequest;
+import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.app.ActivityCompat;
 import androidx.core.content.FileProvider;
-import io.flutter.plugin.common.MethodCall;
-import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry;
+import io.flutter.plugins.imagepicker.Messages.FlutterError;
+import io.flutter.plugins.imagepicker.Messages.ImageSelectionOptions;
+import io.flutter.plugins.imagepicker.Messages.VideoSelectionOptions;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -29,20 +33,14 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-enum CameraDevice {
-  REAR,
-
-  FRONT
-}
-
 /**
  * A delegate class doing the heavy lifting for the plugin.
  *
  * <p>When invoked, both the {@link #chooseImageFromGallery} and {@link #takeImageWithCamera}
  * methods go through the same steps:
  *
- * <p>1. Check for an existing {@link #pendingResult}. If a previous pendingResult exists, this
- * means that the chooseImageFromGallery() or takeImageWithCamera() method was called at least
+ * <p>1. Check for an existing {@link #pendingCallState}. If a previous pendingCallState exists,
+ * this means that the chooseImageFromGallery() or takeImageWithCamera() method was called at least
  * twice. In this case, stop executing and finish with an error.
  *
  * <p>2. Check that a required runtime permission has been granted. The takeImageWithCamera() method
@@ -79,9 +77,31 @@ public class ImagePickerDelegate
   @VisibleForTesting static final int REQUEST_CODE_TAKE_IMAGE_WITH_CAMERA = 2343;
   @VisibleForTesting static final int REQUEST_CAMERA_IMAGE_PERMISSION = 2345;
   @VisibleForTesting static final int REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY = 2346;
+
   @VisibleForTesting static final int REQUEST_CODE_CHOOSE_VIDEO_FROM_GALLERY = 2352;
   @VisibleForTesting static final int REQUEST_CODE_TAKE_VIDEO_WITH_CAMERA = 2353;
   @VisibleForTesting static final int REQUEST_CAMERA_VIDEO_PERMISSION = 2355;
+
+  public enum CameraDevice {
+    REAR,
+    FRONT
+  }
+
+  /** Holds call state during intent handling. */
+  private static class PendingCallState {
+    public final @Nullable ImageSelectionOptions imageOptions;
+    public final @Nullable VideoSelectionOptions videoOptions;
+    public final @NonNull Messages.Result<List<String>> result;
+
+    private PendingCallState(
+        @Nullable ImageSelectionOptions imageOptions,
+        @Nullable VideoSelectionOptions videoOptions,
+        @NonNull Messages.Result<List<String>> result) {
+      this.imageOptions = imageOptions;
+      this.videoOptions = videoOptions;
+      this.result = result;
+    }
+  }
 
   @VisibleForTesting final String fileProviderName;
 
@@ -113,8 +133,7 @@ public class ImagePickerDelegate
   }
 
   private Uri pendingCameraMediaUri;
-  private MethodChannel.Result pendingResult;
-  private MethodCall methodCall;
+  private @Nullable PendingCallState pendingCallState;
 
   public ImagePickerDelegate(
       final Activity activity,
@@ -125,6 +144,7 @@ public class ImagePickerDelegate
         activity,
         externalFilesDirectory,
         imageResizer,
+        null,
         null,
         null,
         cache,
@@ -177,8 +197,9 @@ public class ImagePickerDelegate
       final Activity activity,
       final File externalFilesDirectory,
       final ImageResizer imageResizer,
-      final MethodChannel.Result result,
-      final MethodCall methodCall,
+      final @Nullable ImageSelectionOptions pendingImageOptions,
+      final @Nullable VideoSelectionOptions pendingVideoOptions,
+      final @Nullable Messages.Result<List<String>> result,
       final ImagePickerCache cache,
       final PermissionManager permissionManager,
       final FileUriResolver fileUriResolver,
@@ -187,8 +208,10 @@ public class ImagePickerDelegate
     this.externalFilesDirectory = externalFilesDirectory;
     this.imageResizer = imageResizer;
     this.fileProviderName = activity.getPackageName() + ".flutter.image_provider";
-    this.pendingResult = result;
-    this.methodCall = methodCall;
+    if (result != null) {
+      this.pendingCallState =
+          new PendingCallState(pendingImageOptions, pendingVideoOptions, result);
+    }
     this.permissionManager = permissionManager;
     this.fileUriResolver = fileUriResolver;
     this.fileUtils = fileUtils;
@@ -199,68 +222,91 @@ public class ImagePickerDelegate
     cameraDevice = device;
   }
 
-  CameraDevice getCameraDevice() {
-    return cameraDevice;
-  }
-
   // Save the state of the image picker so it can be retrieved with `retrieveLostImage`.
   void saveStateBeforeResult() {
-    if (methodCall == null) {
+    if (pendingCallState == null) {
       return;
     }
 
-    cache.saveTypeWithMethodCallName(methodCall.method);
-    cache.saveDimensionWithMethodCall(methodCall);
+    cache.saveType(
+        pendingCallState.imageOptions != null
+            ? ImagePickerCache.CacheType.IMAGE
+            : ImagePickerCache.CacheType.VIDEO);
+    if (pendingCallState.imageOptions != null) {
+      cache.saveDimensionWithOutputOptions(pendingCallState.imageOptions);
+    }
     if (pendingCameraMediaUri != null) {
       cache.savePendingCameraMediaUriPath(pendingCameraMediaUri);
     }
   }
 
-  void retrieveLostImage(MethodChannel.Result result) {
-    Map<String, Object> resultMap = cache.getCacheMap();
+  @Nullable
+  Messages.CacheRetrievalResult retrieveLostImage() {
+    Map<String, Object> cacheMap = cache.getCacheMap();
+    if (cacheMap.isEmpty()) {
+      return null;
+    }
+
+    Messages.CacheRetrievalResult.Builder result = new Messages.CacheRetrievalResult.Builder();
+
+    Messages.CacheRetrievalType type =
+        (Messages.CacheRetrievalType) cacheMap.get(ImagePickerCache.MAP_KEY_TYPE);
+    if (type != null) {
+      result.setType(type);
+    }
+    result.setError((Messages.CacheRetrievalError) cacheMap.get(ImagePickerCache.MAP_KEY_ERROR));
     @SuppressWarnings("unchecked")
-    ArrayList<String> pathList = (ArrayList<String>) resultMap.get(cache.MAP_KEY_PATH_LIST);
-    ArrayList<String> newPathList = new ArrayList<>();
+    ArrayList<String> pathList =
+        (ArrayList<String>) cacheMap.get(ImagePickerCache.MAP_KEY_PATH_LIST);
     if (pathList != null) {
+      ArrayList<String> newPathList = new ArrayList<>();
       for (String path : pathList) {
-        Double maxWidth = (Double) resultMap.get(cache.MAP_KEY_MAX_WIDTH);
-        Double maxHeight = (Double) resultMap.get(cache.MAP_KEY_MAX_HEIGHT);
-        int imageQuality =
-            resultMap.get(cache.MAP_KEY_IMAGE_QUALITY) == null
-                ? 100
-                : (int) resultMap.get(cache.MAP_KEY_IMAGE_QUALITY);
+        Double maxWidth = (Double) cacheMap.get(ImagePickerCache.MAP_KEY_MAX_WIDTH);
+        Double maxHeight = (Double) cacheMap.get(ImagePickerCache.MAP_KEY_MAX_HEIGHT);
+        Integer boxedImageQuality = (Integer) cacheMap.get(ImagePickerCache.MAP_KEY_IMAGE_QUALITY);
+        int imageQuality = boxedImageQuality == null ? 100 : boxedImageQuality;
 
         newPathList.add(imageResizer.resizeImageIfNeeded(path, maxWidth, maxHeight, imageQuality));
       }
-      resultMap.put(cache.MAP_KEY_PATH_LIST, newPathList);
-      resultMap.put(cache.MAP_KEY_PATH, newPathList.get(newPathList.size() - 1));
+      result.setPaths(newPathList);
     }
-    if (resultMap.isEmpty()) {
-      result.success(null);
-    } else {
-      result.success(resultMap);
-    }
+
     cache.clear();
+
+    return result.build();
   }
 
-  public void chooseVideoFromGallery(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void chooseVideoFromGallery(
+      VideoSelectionOptions options, boolean usePhotoPicker, Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(null, options, result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
 
-    launchPickVideoFromGalleryIntent();
+    launchPickVideoFromGalleryIntent(usePhotoPicker);
   }
 
-  private void launchPickVideoFromGalleryIntent() {
-    Intent pickVideoIntent = new Intent(Intent.ACTION_GET_CONTENT);
-    pickVideoIntent.setType("video/*");
+  private void launchPickVideoFromGalleryIntent(Boolean useAndroidPhotoPicker) {
+    Intent pickVideoIntent;
+    if (useAndroidPhotoPicker && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+      pickVideoIntent =
+          new ActivityResultContracts.PickVisualMedia()
+              .createIntent(
+                  activity,
+                  new PickVisualMediaRequest.Builder()
+                      .setMediaType(ActivityResultContracts.PickVisualMedia.VideoOnly.INSTANCE)
+                      .build());
+    } else {
+      pickVideoIntent = new Intent(Intent.ACTION_GET_CONTENT);
+      pickVideoIntent.setType("video/*");
+    }
 
     activity.startActivityForResult(pickVideoIntent, REQUEST_CODE_CHOOSE_VIDEO_FROM_GALLERY);
   }
 
-  public void takeVideoWithCamera(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void takeVideoWithCamera(
+      VideoSelectionOptions options, Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(null, options, result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
@@ -277,8 +323,10 @@ public class ImagePickerDelegate
 
   private void launchTakeVideoWithCameraIntent() {
     Intent intent = new Intent(MediaStore.ACTION_VIDEO_CAPTURE);
-    if (this.methodCall != null && this.methodCall.argument("maxDuration") != null) {
-      int maxSeconds = this.methodCall.argument("maxDuration");
+    if (pendingCallState != null
+        && pendingCallState.videoOptions != null
+        && pendingCallState.videoOptions.getMaxDurationSeconds() != null) {
+      int maxSeconds = pendingCallState.videoOptions.getMaxDurationSeconds().intValue();
       intent.putExtra(MediaStore.EXTRA_DURATION_LIMIT, maxSeconds);
     }
     if (cameraDevice == CameraDevice.FRONT) {
@@ -306,43 +354,72 @@ public class ImagePickerDelegate
     }
   }
 
-  public void chooseImageFromGallery(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void chooseImageFromGallery(
+      @NonNull ImageSelectionOptions options,
+      boolean usePhotoPicker,
+      Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(options, null, result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
 
-    launchPickImageFromGalleryIntent();
+    launchPickImageFromGalleryIntent(usePhotoPicker);
   }
 
-  public void chooseMultiImageFromGallery(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void chooseMultiImageFromGallery(
+      @NonNull ImageSelectionOptions options,
+      boolean usePhotoPicker,
+      Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(options, null, result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
 
-    launchMultiPickImageFromGalleryIntent();
+    launchMultiPickImageFromGalleryIntent(usePhotoPicker);
   }
 
-  private void launchPickImageFromGalleryIntent() {
-    Intent pickImageIntent = new Intent(Intent.ACTION_GET_CONTENT);
-    pickImageIntent.setType("image/*");
+  private void launchPickImageFromGalleryIntent(Boolean useAndroidPhotoPicker) {
+    Intent pickImageIntent;
+    if (useAndroidPhotoPicker && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+      pickImageIntent =
+          new ActivityResultContracts.PickVisualMedia()
+              .createIntent(
+                  activity,
+                  new PickVisualMediaRequest.Builder()
+                      .setMediaType(ActivityResultContracts.PickVisualMedia.ImageOnly.INSTANCE)
+                      .build());
+    } else {
+      pickImageIntent = new Intent(Intent.ACTION_GET_CONTENT);
+      pickImageIntent.setType("image/*");
+    }
 
     activity.startActivityForResult(pickImageIntent, REQUEST_CODE_CHOOSE_IMAGE_FROM_GALLERY);
   }
 
-  private void launchMultiPickImageFromGalleryIntent() {
-    Intent pickImageIntent = new Intent(Intent.ACTION_GET_CONTENT);
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
-      pickImageIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
+  private void launchMultiPickImageFromGalleryIntent(Boolean useAndroidPhotoPicker) {
+    Intent pickMultiImageIntent;
+    if (useAndroidPhotoPicker && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+      pickMultiImageIntent =
+          new ActivityResultContracts.PickMultipleVisualMedia()
+              .createIntent(
+                  activity,
+                  new PickVisualMediaRequest.Builder()
+                      .setMediaType(ActivityResultContracts.PickVisualMedia.ImageOnly.INSTANCE)
+                      .build());
+    } else {
+      pickMultiImageIntent = new Intent(Intent.ACTION_GET_CONTENT);
+      pickMultiImageIntent.setType("image/*");
+      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
+        pickMultiImageIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
+      }
     }
-    pickImageIntent.setType("image/*");
-
-    activity.startActivityForResult(pickImageIntent, REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY);
+    activity.startActivityForResult(
+        pickMultiImageIntent, REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY);
   }
 
-  public void takeImageWithCamera(MethodCall methodCall, MethodChannel.Result result) {
-    if (!setPendingMethodCallAndResult(methodCall, result)) {
+  public void takeImageWithCamera(
+      @NonNull ImageSelectionOptions options, Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(options, null, result)) {
       finishWithAlreadyActiveError(result);
       return;
     }
@@ -414,6 +491,8 @@ public class ImagePickerDelegate
 
   private void grantUriPermissions(Intent intent, Uri imageUri) {
     PackageManager packageManager = activity.getPackageManager();
+    // TODO(stuartmorgan): Add new codepath: https://github.com/flutter/flutter/issues/121816
+    @SuppressWarnings("deprecation")
     List<ResolveInfo> compatibleActivities =
         packageManager.queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY);
 
@@ -427,7 +506,7 @@ public class ImagePickerDelegate
 
   @Override
   public boolean onRequestPermissionsResult(
-      int requestCode, String[] permissions, int[] grantResults) {
+      int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
     boolean permissionGranted =
         grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED;
 
@@ -563,10 +642,10 @@ public class ImagePickerDelegate
 
   private void handleMultiImageResult(
       ArrayList<String> paths, boolean shouldDeleteOriginalIfScaled) {
-    if (methodCall != null) {
+    if (pendingCallState != null && pendingCallState.imageOptions != null) {
       ArrayList<String> finalPath = new ArrayList<>();
       for (int i = 0; i < paths.size(); i++) {
-        String finalImagePath = getResizedImagePath(paths.get(i));
+        String finalImagePath = getResizedImagePath(paths.get(i), pendingCallState.imageOptions);
 
         //delete original file if scaled
         if (finalImagePath != null
@@ -583,8 +662,8 @@ public class ImagePickerDelegate
   }
 
   private void handleImageResult(String path, boolean shouldDeleteOriginalIfScaled) {
-    if (methodCall != null) {
-      String finalImagePath = getResizedImagePath(path);
+    if (pendingCallState != null && pendingCallState.imageOptions != null) {
+      String finalImagePath = getResizedImagePath(path, pendingCallState.imageOptions);
       //delete original file if scaled
       if (finalImagePath != null && !finalImagePath.equals(path) && shouldDeleteOriginalIfScaled) {
         new File(path).delete();
@@ -595,26 +674,27 @@ public class ImagePickerDelegate
     }
   }
 
-  private String getResizedImagePath(String path) {
-    Double maxWidth = methodCall.argument("maxWidth");
-    Double maxHeight = methodCall.argument("maxHeight");
-    Integer imageQuality = methodCall.argument("imageQuality");
-
-    return imageResizer.resizeImageIfNeeded(path, maxWidth, maxHeight, imageQuality);
+  private String getResizedImagePath(String path, @NonNull ImageSelectionOptions outputOptions) {
+    return imageResizer.resizeImageIfNeeded(
+        path,
+        outputOptions.getMaxWidth(),
+        outputOptions.getMaxHeight(),
+        outputOptions.getQuality().intValue());
   }
 
   private void handleVideoResult(String path) {
     finishWithSuccess(path);
   }
 
-  private boolean setPendingMethodCallAndResult(
-      MethodCall methodCall, MethodChannel.Result result) {
-    if (pendingResult != null) {
+  private boolean setPendingOptionsAndResult(
+      @Nullable ImageSelectionOptions imageOptions,
+      @Nullable VideoSelectionOptions videoOptions,
+      @NonNull Messages.Result<List<String>> result) {
+    if (pendingCallState != null) {
       return false;
     }
 
-    this.methodCall = methodCall;
-    pendingResult = result;
+    pendingCallState = new PendingCallState(imageOptions, videoOptions, result);
 
     // Clean up cache if a new image picker is launched.
     cache.clear();
@@ -627,44 +707,41 @@ public class ImagePickerDelegate
   // A null imagePath indicates that the image picker was cancelled without
   // selection.
   private void finishWithSuccess(@Nullable String imagePath) {
-    if (pendingResult == null) {
+    ArrayList<String> pathList = new ArrayList<>();
+    if (imagePath != null) {
+      pathList.add(imagePath);
+    }
+    if (pendingCallState == null) {
       // Only save data for later retrieval if something was actually selected.
-      if (imagePath != null) {
-        ArrayList<String> pathList = new ArrayList<>();
-        pathList.add(imagePath);
+      if (!pathList.isEmpty()) {
         cache.saveResult(pathList, null, null);
       }
       return;
     }
-    pendingResult.success(imagePath);
-    clearMethodCallAndResult();
+    pendingCallState.result.success(pathList);
+    pendingCallState = null;
   }
 
   private void finishWithListSuccess(ArrayList<String> imagePaths) {
-    if (pendingResult == null) {
+    if (pendingCallState == null) {
       cache.saveResult(imagePaths, null, null);
       return;
     }
-    pendingResult.success(imagePaths);
-    clearMethodCallAndResult();
+    pendingCallState.result.success(imagePaths);
+    pendingCallState = null;
   }
 
-  private void finishWithAlreadyActiveError(MethodChannel.Result result) {
-    result.error("already_active", "Image picker is already active", null);
+  private void finishWithAlreadyActiveError(Messages.Result<List<String>> result) {
+    result.error(new FlutterError("already_active", "Image picker is already active", null));
   }
 
   private void finishWithError(String errorCode, String errorMessage) {
-    if (pendingResult == null) {
+    if (pendingCallState == null) {
       cache.saveResult(null, errorCode, errorMessage);
       return;
     }
-    pendingResult.error(errorCode, errorMessage, null);
-    clearMethodCallAndResult();
-  }
-
-  private void clearMethodCallAndResult() {
-    methodCall = null;
-    pendingResult = null;
+    pendingCallState.result.error(new FlutterError(errorCode, errorMessage, null));
+    pendingCallState = null;
   }
 
   private void useFrontCamera(Intent intent) {