+++ packages/local_auth/local_auth_ios/CHANGELOG.md
@@ -1,5 +1,15 @@
-## NEXT
+## 1.1.1
 
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
+## 1.1.0
+
+* Updates minimum Flutter version to 3.3 and iOS 11.
+
+## 1.0.13
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 1.0.12