+++ packages/shared_preferences/shared_preferences_windows/CHANGELOG.md
@@ -1,5 +1,11 @@
-## NEXT
+## 2.1.5
 
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
+## 2.1.4
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.1.3