+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -1,5 +1,19 @@
-## NEXT
+## 2.4.2
 
+* Makes seekTo async and only complete when AVPlayer.seekTo completes.
+
+## 2.4.1
+
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
+## 2.4.0
+
+* Updates minimum Flutter version to 3.3 and iOS 11.
+
+## 2.3.9
+
+* Updates links for the merge of flutter/plugins into flutter/packages.
 * Updates minimum Flutter version to 3.0.
 
 ## 2.3.8